fix code PostManagement và add code Phương
